refactored(search): moved search results filter translation into enum
